chore: remove logotype svg annd move logotype_text
