+++ lib/components/elements/Icon/logotype_text.js
