+++ lib/components/wizard/GeneratingKeys.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Paragraph } from '../typography/Typography'
 import { Spinner } from '../elements/Spinner/Spinner'
-import Logo from './logotype'
+import Logo from '../elements/Icon/logotype_text'
 
 function GeneratingKeys({ status }) {
   return (
